requirements.txt
langchain~=0.3.27
langchain-openai~=0.3.32
langsmith
langchain-experimental
python-dotenv~=1.1.1
langchain-community~=0.3.28
faiss-cpu==1.8.0
langchain-text-splitters~=0.3.9
langchain-core~=0.3.75